index.js: fix lens position updating, refactor code
Trying to separately track the current and target position is futile
when the lens can be moved independently of HomeKit. Instead, let's
track only the lens' actual position and report this to HomeKit as both
current and target positions. This simplifies the logic and resolves
all various issues when current and target are out of sync.

Separately, refactor and clean up the code a bit.